split up interpolatePath()
