+++ packages/paths/src/test/interpolate-path.test.ts
@@ -1,5 +1,5 @@
 import {expect, test} from 'bun:test';
-import {interpolatePath} from '../interpolate-path';
+import {interpolatePath} from '../interpolate-path/interpolate-path';
 
 test('Should be able to interpolate path', () => {
 	expect(interpolatePath(0.5, 'M 0 0 L 100 0', 'M 100 0 L 0 0')).toBe(
